refactor: rename some files to .ts
