+++ src/Widget/Clock/index.ts
