+++ src/Widget/Note/index.ts
