+++ src/Widget/Weather/index.ts
