+++ src/Widget/widgetConfig.ts
