+++ src/components/RoundButton/index.ts
@@ -1,0 +1,4 @@
+import RoundLinkButton from './RoundLinkButton';
+import RoundAddButton from './RoundAddButton';
+
+export { RoundLinkButton, RoundAddButton };
