--- src/components/RoundButton/index.tsx
@@ -1,4 +1,0 @@
-import RoundLinkButtonProps from './RoundLinkButtonProps';
-import RoundAddButton from './RoundAddButton';
-
-export { RoundLinkButtonProps, RoundAddButton };
